refactor: remove dependency on swagger in common
(cherry picked from commit da7a912a5cdf49eecac97a6195deec0727143fd7)
