--- app/common/model/src/main/resources/META-INF/services/io.swagger.v3.core.converter.ModelConverter
@@ -1,1 +1,0 @@
-io.syndesis.common.model.support.swagger.ModelConverter
