+++ app/server/runtime/src/main/java/io/syndesis/server/runtime/swagger/ModelConverter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.syndesis.common.model.support.swagger;
+package io.syndesis.server.runtime.swagger;
 
 import java.util.Set;
