+++ app/server/runtime/src/main/resources/META-INF/services/io.swagger.v3.core.converter.ModelConverter
@@ -1,0 +1,1 @@
+io.syndesis.server.runtime.swagger.ModelConverter
