+++ app/server/runtime/src/test/java/io/syndesis/server/runtime/swagger/ModelConverterTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.syndesis.common.model.support.swagger;
+package io.syndesis.server.runtime.swagger;
 
 import java.lang.reflect.Method;
 import java.util.Collections;
